+++ Hearthstone Deck Tracker/Hearthstone/EffectSystem/Effects/Druid/ConstructPylonsEnchantment.cs
@@ -1,0 +1,16 @@
+﻿using Hearthstone_Deck_Tracker.Hearthstone.EffectSystem.Enums;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.EffectSystem.Effects.Druid;
+
+public class ConstructPylonsEnchantment : EntityBasedEffect
+{
+	public override string CardId => HearthDb.CardIds.NonCollectible.Druid.ConstructPylons_PsionicMatrixEnchantment;
+	protected override string CardIdToShowInUI => HearthDb.CardIds.Collectible.Druid.ConstructPylons;
+
+	public ConstructPylonsEnchantment(int entityId, bool isControlledByPlayer) : base(entityId, isControlledByPlayer)
+	{
+	}
+
+	public override EffectDuration EffectDuration => EffectDuration.Conditional;
+	public override EffectTag EffectTag => EffectTag.CostModification;
+}