+++ Hearthstone Deck Tracker/Hearthstone/EffectSystem/Effects/Mage/ShieldBatteryEnchantment.cs
@@ -1,0 +1,16 @@
+﻿using Hearthstone_Deck_Tracker.Hearthstone.EffectSystem.Enums;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.EffectSystem.Effects.Mage;
+
+public class ShieldBatteryEnchantment : EntityBasedEffect
+{
+	public override string CardId => HearthDb.CardIds.NonCollectible.Neutral.ShieldBattery_KhalaiIngenuityEnchantment;
+	protected override string CardIdToShowInUI => HearthDb.CardIds.Collectible.Mage.ShieldBattery;
+
+	public ShieldBatteryEnchantment(int entityId, bool isControlledByPlayer) : base(entityId, isControlledByPlayer)
+	{
+	}
+
+	public override EffectDuration EffectDuration => EffectDuration.Conditional;
+	public override EffectTag EffectTag => EffectTag.CostModification;
+}