+++ Hearthstone Deck Tracker/Hearthstone/EffectSystem/Effects/Neutral/WarpGateEnchantment.cs
@@ -1,0 +1,16 @@
+﻿using Hearthstone_Deck_Tracker.Hearthstone.EffectSystem.Enums;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.EffectSystem.Effects.Neutral;
+
+public class WarpGateEnchantment : EntityBasedEffect
+{
+	public override string CardId => HearthDb.CardIds.NonCollectible.Neutral.WarpGate_WarpConduitEnchantment;
+	protected override string CardIdToShowInUI => HearthDb.CardIds.Collectible.Invalid.WarpGate;
+
+	public WarpGateEnchantment(int entityId, bool isControlledByPlayer) : base(entityId, isControlledByPlayer)
+	{
+	}
+
+	public override EffectDuration EffectDuration => EffectDuration.Conditional;
+	public override EffectTag EffectTag => EffectTag.CostModification;
+}